Struct indexable_str::IndexableStr

pub struct IndexableStr<'a> { /* private fields */ }

IndexableStr is a struct for creating immutable string objects that make text parsing with Rust a bit more elegant.

IndexableStr can be used to retrieve a char from a specified index as follows:

use indexable_str::IndexableStr;
 
let s = IndexableStr::new("0😀2345678😀");
 
assert_eq!(s[1], '😀');

IndexableStr also allows creating strs over a range of chars as follows:

use indexable_str::IndexableStr;
 
let s = IndexableStr::new("0😀2345678😀");
 
assert_eq!(&s[1..9], "😀2345678")

IndexableStr is designed to work well with all valid UTF-8 characters.

You should note that IndexableStr creates a vector of objects that holds a char and the starting byte offset of the char’s position in the underlying string as a usize. This requires additional memory resources. However, the convenience of IndexableStr should outweigh the additional memory requirements for most applications.

Implementations

impl<'a> IndexableStr<'a>

pub fn new(str: &'a str) -> IndexableStr<'_>

Parameters:

str: &'a str is the string to be made indexable.

Returns an IndexableStr object for the specified str argument.

use indexable_str::IndexableStr;
 
let s = IndexableStr::new("0😀2345678😀");

pub fn as_str(&self) -> &'a str

Returns an &'a str for the underlying string as follows.

use indexable_str::IndexableStr;
 
let s = IndexableStr::new("0😀2345678😀");
 
assert_eq!(s.as_str(), "0😀2345678😀");

pub fn len(&self) -> usize

Returns a usize for the number of chars in the string.

use indexable_str::IndexableStr;
 
let s = IndexableStr::new("0😀2345678😀");
 
assert_eq!(s.len(), 10);